+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Maintenance:
 * `symfony/monolog-bundle` upgraded to ^4.0; review your monolog configuration if you have customised it outside of the defaults.
 
 Changes:
+* The metadata expiration time (`validUntil` attribute) is now configurable via the `metadata_expiration_time` parameter in `parameters.yml` (default: 86400 seconds / 24 hours).
 * The `consent.deleted_at` should be not nullable, and have a default value of `0000-00-00 00:00:00`.
   * Because `deleted_at` is part of the PK, no migration is provided. The database engine should not allow this to be null in the first place, so it is probably not nullable already on your db.
   * The `0000-00-00 00:00:00` is added for clarity/consistency, as this is probably the default behaviour of your database already.